requirements.txt
# Core Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.22.0

# Data Validation and Settings
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Environment Variables
python-dotenv>=1.0.0

# HTTP Requests
requests>=2.31.0

# Date/Time Utilities
python-dateutil>=2.8.2

# XML Processing (for SOAP)
xmltodict>=0.13.0
lxml>=4.9.3
zeep>=4.2.1

# Database Connectivity
mysql-connector-python>=8.0.33

# Excel Processing
openpyxl>=3.1.2
pandas>=2.0.0

# Additional Dependencies
python-multipart>=0.0.6  # For file uploads in FastAPI

# Monitoring and Observability
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-instrumentation-requests>=0.41b0
opentelemetry-exporter-otlp>=1.20.0
prometheus-client>=0.18.0